requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
tensorflow==2.13.0
plotly==5.16.1
requests==2.32.3
urllib3==2.4.0
certifi>=2025.4.26
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.0.0 